fix: langchain
