+++ packages/opentelemetry-instrumentation-langchain/tests/test_lcel.py
@@ -1,0 +1,27 @@
+import pytest
+from langchain.prompts import PromptTemplate
+from langchain.schema import StrOutputParser
+from langchain.chat_models import ChatOpenAI
+
+
+@pytest.mark.vcr
+def test_langchain_streaming(exporter):
+    chat = ChatOpenAI(
+        model="gpt-4",
+        temperature=0,
+        streaming=True,
+    )
+
+    prompt = PromptTemplate.from_template(
+        "write 10 lines of random text about ${product}"
+    )
+    runnable = prompt | chat | StrOutputParser()
+    runnable.invoke({"product": "colorful socks"})
+
+    spans = exporter.get_finished_spans()
+
+    assert set(
+        [
+            "openai.chat",
+        ]
+    ).issubset([span.name for span in spans])
